requirements.txt
Django==1.9.7
django-autoslug==1.9.3
django-cors-headers==1.1.0
djangorestframework==3.3.3
markdown2==2.3.1
mysql-connector==2.1.3
mysqlclient==1.3.7
